+++ src/handlers/private/mod.rs
@@ -1,1 +1,2 @@
+pub mod mqtt_auth;
 pub mod offline_message;