+++ src/models/api/mod.rs
@@ -1,5 +1,6 @@
 pub mod auth;
 pub mod bundle;
 pub mod device;
+pub mod mqtt;
 pub mod user;
 pub mod webhook;